+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/Rules.hs
@@ -4,6 +4,7 @@
 
 module Cardano.Ledger.Babbage.Rules (
   module Cardano.Ledger.Babbage.Rules.Ledger,
+  module Cardano.Ledger.Babbage.Rules.Ledgers,
   module Cardano.Ledger.Babbage.Rules.Utxo,
   module Cardano.Ledger.Babbage.Rules.Utxos,
   module Cardano.Ledger.Babbage.Rules.Utxow,
@@ -16,7 +17,7 @@ import Cardano.Ledger.Babbage.Rules.Deleg ()
 import Cardano.Ledger.Babbage.Rules.Delegs ()
 import Cardano.Ledger.Babbage.Rules.Delpl ()
 import Cardano.Ledger.Babbage.Rules.Ledger
-import Cardano.Ledger.Babbage.Rules.Ledgers ()
+import Cardano.Ledger.Babbage.Rules.Ledgers
 import Cardano.Ledger.Babbage.Rules.Pool ()
 import Cardano.Ledger.Babbage.Rules.Ppup ()
 import Cardano.Ledger.Babbage.Rules.Utxo